Replaces curl with wget for AMP installation
Uses wget instead of curl in the AMP installation script.

This change aims to improve the reliability of the installation process,
as wget might be more readily available or perform better in certain
environments.

Replaces curl with wget for AMP installation

Uses wget instead of curl in the AMP installation script.

This change aims to improve the reliability of the
installation process, as wget might be more readily
available or perform better in certain environments.